+++ src/components/AccordionPanel/AccordionPanelTrigger.tsx
@@ -1,7 +1,7 @@
-import React, { FC, useContext, useCallback } from 'react'
+import React, { FC, useCallback, useContext } from 'react'
 import styled, { css } from 'styled-components'
 
-import { useTheme, Theme } from '../../hooks/useTheme'
+import { Theme, useTheme } from '../../hooks/useTheme'
 import { getIsInclude, mapToKeyArray } from '../../libs/map'
 import { isTouchDevice } from '../../libs/ua'
 import { getNewExpandedItems } from './accordionPanelHelper'