+++ src/components/Dialog/MessageDialog.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useEffect } from 'react'
+import React, { useEffect, useRef } from 'react'
 import { createPortal } from 'react-dom'
 
 import { DialogContentInner } from './DialogContentInner'
@@ -18,8 +18,12 @@ export const MessageDialog: React.FC<Props> = ({
   description,
   closeText,
   onClickClose,
-  onClickOverlay = () => {},
-  onPressEscape = () => {},
+  onClickOverlay = () => {
+    /* noop */
+  },
+  onPressEscape = () => {
+    /* noop */
+  },
   ...props
 }) => {
   const element = useRef(document.createElement('div')).current