+++ src/components/Dropdown/DropdownTrigger.tsx
@@ -9,11 +9,12 @@ type Props = {
 }
 
 export const DropdownTrigger: React.FC<Props> = ({ children, className = '' }) => {
-  const { dropdownKey, active, onClickTrigger } = useContext(DropdownContext)
+  const { active, onClickTrigger, triggerElementRef } = useContext(DropdownContext)
 
   return (
     <Wrapper
-      onClick={e => {
+      ref={triggerElementRef}
+      onClick={(e) => {
         const rect = e.currentTarget.getBoundingClientRect()
         onClickTrigger({
           top: rect.top,
@@ -22,7 +23,7 @@ export const DropdownTrigger: React.FC<Props> = ({ children, className = '' }) =
           left: rect.left,
         })
       }}
-      className={`${dropdownKey} ${className}`}
+      className={className}
     >
       {React.Children.map(children, (child: any) => {
         const props = child.props ? child.props : {}