+++ src/components/Select/Select.tsx
@@ -1,8 +1,8 @@
-import React, { SelectHTMLAttributes, FC, ChangeEvent, useCallback } from 'react'
+import React, { ChangeEvent, FC, SelectHTMLAttributes, useCallback } from 'react'
 import styled, { css } from 'styled-components'
 
 import { isTouchDevice } from '../../libs/ua'
-import { useTheme, Theme } from '../../hooks/useTheme'
+import { Theme, useTheme } from '../../hooks/useTheme'
 
 import { Icon } from '../Icon'
 
@@ -41,7 +41,7 @@ export const Select: FC<Props> = ({
   return (
     <Wrapper className={className} width={widthStyle} theme={theme}>
       <SelectBox className={error ? 'error' : ''} onChange={handleChange} themes={theme} {...props}>
-        {options.map(option => {
+        {options.map((option) => {
           if ('value' in option) {
             return (
               <option key={option.value} value={option.value}>
@@ -54,7 +54,7 @@ export const Select: FC<Props> = ({
 
           return (
             <optgroup key={optgroup.label} label={optgroup.label}>
-              {optgroup.options.map(o => (
+              {optgroup.options.map((o) => (
                 <option key={o.value} value={o.value}>
                   {o.label}
                 </option>